Migrate ModelExpress metadata publishing to RemoteInstanceWeightTransport
